+++ R/chrono.subsets_fun.R
@@ -5,6 +5,10 @@ get.percent.age <- function(tree, percent = 0.01) {
     while(is.null(tree_slice) || Ntip(tree_slice) < 3) {
         percent <- percent + 0.01
         tree_slice <- slice.tree.sharp(tree, tree$root.time - (percent * tree$root.time))
+        if(percent >= 100) {
+            stop("Impossible to find a starting point to slice the tree. This can happen if the tree has no branch length or has a \"ladder\" structure. You can try to fix that by setting specific slicing times.")
+            break
+        }
     }
     return(percent)
 }