+++ tests/testthat/test-chrono.subsets.R
@@ -785,4 +785,16 @@ test_that("fast internal functions work", {
     expect_equal(res2[[1]][,3], c(1,1,.5,1))
 })
 
+test_that("infinite loop blocker for get.percent.age", {
+   ## Testing data
+    matrix <- do.call(rbind, list(matrix(1, 5, 5), matrix(2, 3, 5), matrix(3, 4, 5)))
+    rownames(matrix) <- paste0("t", 1:12)
+    test_tree <- stree(12, type = "right")
+    test_tree$edge.length <- rep(1, Nedge(test_tree))
+    test_tree$root.time <- 12
+
+    ## custom subsets
+    error <- capture_error(chrono.subsets(matrix, test_tree, method = "continuous", time = 5, model = "acctran"))
+    expect_equal(error[[1]], "Impossible to find a starting point to slice the tree. This can happen if the tree has no branch length or has a \"ladder\" structure. You can try to fix that by setting specific slicing times.")
+})